Move GetDepositsResponse to Teku specific location
